+++ packages/flame/lib/experimental.dart
@@ -9,12 +9,20 @@
 /// After the components lived here for some time, and when we gain more
 /// confidence in their robustness, they will be moved out into the main Flame
 /// library.
-
 export 'src/experimental/camera_component.dart' show CameraComponent;
 export 'src/experimental/circular_viewport.dart' show CircularViewport;
 export 'src/experimental/fixed_aspect_ratio_viewport.dart'
     show FixedAspectRatioViewport;
 export 'src/experimental/fixed_size_viewport.dart' show FixedSizeViewport;
+export 'src/experimental/geometry/components/shape_component.dart'
+    show ShapeComponent;
+export 'src/experimental/geometry/shapes/circle.dart' show Circle;
+export 'src/experimental/geometry/shapes/line_segment.dart' show LineSegment;
+export 'src/experimental/geometry/shapes/polygon.dart' show Polygon;
+export 'src/experimental/geometry/shapes/rectangle.dart' show Rectangle;
+export 'src/experimental/geometry/shapes/rounded_rectangle.dart'
+    show RoundedRectangle;
+export 'src/experimental/geometry/shapes/shape.dart' show Shape;
 export 'src/experimental/max_viewport.dart' show MaxViewport;
 export 'src/experimental/viewfinder.dart' show Viewfinder;
 export 'src/experimental/viewport.dart' show Viewport;
