+++ packages/flame/lib/src/experimental/geometry/shapes/polygon.dart
@@ -2,7 +2,7 @@ import 'dart:ui';
 
 import 'package:vector_math/vector_math_64.dart';
 
-import '../../game/transform2d.dart';
+import '../../../game/transform2d.dart';
 import 'shape.dart';
 
 class Polygon extends Shape {
